Pulled out RigidBodyActuator into its own files. (#2889)
* Pulled out RigidBodyActuator into its own files.

* Ran clang-format.

* Modified to use forward declaration.

* Added missing #include.
